n_poly

+++ src/n_poly/n_polyu1n_gcd.c
@@ -409,11 +409,11 @@ choose_prime: /* primes are v - alpha, v + alpha */
     if (n_poly_degree(modulus) > 0)
     {
         FLINT_ASSERT(G->length > 0);
-        if (n_poly_degree(Gevalp) > G->exps[0])
+        if ((ulong) n_poly_degree(Gevalp) > G->exps[0])
         {
             goto choose_prime;
         }
-        else if (n_poly_degree(Gevalp) < G->exps[0])
+        else if ((ulong) n_poly_degree(Gevalp) < G->exps[0])
         {
             n_poly_one(modulus);
         }
